+++ src/app/suppliers/page.tsx
@@ -1,3 +1,4 @@
+"use client";
 import { NoAsset } from "~/components/Assets";
 import Layout from "~/components/Layout/Layout";
 import SupplierList from "~/components/suppliers/SupplierList";
@@ -18,7 +19,7 @@ export default function Page() {
     <Layout>
       <main className="pl-5">
         {isError && <h3>We have an error</h3>}
-        <div className="flex md:w-[1000px] items-center justify-between pt-[40px] ">
+        <div className="flex items-center justify-between pt-[40px] md:w-[1000px] ">
           <h3 className="text-3xl font-medium ">Your Company </h3>
           <div className="flex items-center gap-2">
             <Link href="/suppliers/">
@@ -44,7 +45,9 @@ export default function Page() {
           </h3>
         )}
 
-        {data?.length !== 0 &&  data !==  null && !isLoading && <SupplierList suppliers={data} />}
+        {data?.length !== 0 && data !== null && !isLoading && (
+          <SupplierList suppliers={data} />
+        )}
         {isLoading && <LoadingSkeleton />}
       </main>
     </Layout>
